comments and style improvements
