+++ src/main/java/com/about/java/controllers/UploadServlet.java
@@ -4,7 +4,7 @@ package com.about.java.controllers;
  * Created by Юыќур on 05.11.2015.
  */
 
-import FileRepository.FileRepository;
+import Repository.FileRepository;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.MultipartConfig;
@@ -18,53 +18,53 @@ import java.io.IOException;
 
 @WebServlet("/UploadServlet")
 @MultipartConfig(fileSizeThreshold = 1024 * 1024 * 2, // 2MB
-        maxFileSize = 1024 * 1024 * 10,      // 10MB
-        maxRequestSize = 1024 * 1024 * 50)   // 50MB
+		maxFileSize = 1024 * 1024 * 10,      // 10MB
+		maxRequestSize = 1024 * 1024 * 50)   // 50MB
 public class UploadServlet extends HttpServlet {
-    /**
-     * Name of the directory where uploaded files will be saved, relative to
-     * the web application directory.
-     */
-    private static final String SAVE_DIR = "uploadFiles";
-    /**
-     * handles file upload
-     */
+	/**
+	 * Name of the directory where uploaded files will be saved, relative to
+	 * the web application directory.
+	 */
+	private static final String saveDir = "uploadFiles";
 
-    protected void doPost(HttpServletRequest request,
-                          HttpServletResponse response) throws ServletException, IOException {
-        // gets absolute path of the web application
-        String appPath = request.getServletContext().getRealPath("");
-        // constructs path of the directory to save uploaded file
-        String savePath = appPath + File.separator + SAVE_DIR;
-        if(!FileRepository.repo.isRootFolderInitialized())
-        {
-            FileRepository.repo.createRootFolder(savePath);
-        }
-        try {
-            String fileName = "";
-            for (Part part : request.getParts()) {
-                fileName = extractFileName(part);
-                FileRepository.repo.addNewFile(part, fileName, "guest");
-            }
-            response.setCharacterEncoding("UTF-32");
-            response.getWriter().write("File " + fileName + " successfully uploaded");
-        } catch (Exception ex) {
-            response.getWriter().write("There was an error, please try again ");
-            response.getWriter().write("Exception info: " + ex.getMessage());
-        }
-    }
+	/**
+	 * handles file upload
+	 */
 
-    /**
-     * Extracts file name from HTTP header content-disposition
-     */
-    private String extractFileName(Part part) {
-        String contentDisp = part.getHeader("content-disposition");
-        String[] items = contentDisp.split(";");
-        for (String s : items) {
-            if (s.trim().startsWith("filename")) {
-                return s.substring(s.indexOf("=") + 2, s.length() - 1);
-            }
-        }
-        return "";
-    }
+	protected void doPost(HttpServletRequest request,
+						  HttpServletResponse response) throws ServletException, IOException {
+		// gets absolute path of the web application
+		String appPath = request.getServletContext().getRealPath("");
+		// constructs path of the directory to save uploaded file
+		String savePath = appPath + File.separator + saveDir;
+		if (!FileRepository.repo.isRootFolderInitialized()) {
+			FileRepository.repo.createRootFolder(savePath);
+		}
+		try {
+			String fileName = "";
+			for (Part part : request.getParts()) {
+				fileName = extractFileName(part);
+				Repository.FileInfo fileInRepo = FileRepository.repo.addNewFile(part, fileName, "guest");
+			}
+			response.setCharacterEncoding("UTF-32");
+			response.getWriter().write("File " + fileName + " successfully uploaded");
+		} catch (Exception ex) {
+			response.getWriter().write("There was an error, please try again ");
+			response.getWriter().write("Exception info: " + ex.getMessage());
+		}
+	}
+
+	/**
+	 * Extracts file name from HTTP header content-disposition
+	 */
+	private String extractFileName(Part part) {
+		String contentDisp = part.getHeader("content-disposition");
+		String[] items = contentDisp.split(";");
+		for (String s : items) {
+			if (s.trim().startsWith("filename")) {
+				return s.substring(s.indexOf("=") + 2, s.length() - 1);
+			}
+		}
+		return "";
+	}
 }
